pyproject.toml
[tool.poetry]
name = "llm-code-refinery"
version = "0.1.0"
description = ""
authors = ["este6an13 <diquintero@unal.edu.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.65.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
